DTSCCI-1461 move notifyParties implementation to process-specific notifier implementation
