+++ src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/Notifier.java
@@ -10,7 +10,6 @@ import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
 import uk.gov.hmcts.reform.civil.service.OrganisationService;
 import uk.gov.hmcts.reform.civil.service.flowstate.SimpleStateFlowEngine;
 
-import java.util.HashSet;
 import java.util.Set;
 
 @Component
@@ -22,21 +21,14 @@ public abstract class Notifier implements NotificationData {
     protected final OrganisationService organisationService;
     protected final SimpleStateFlowEngine stateFlowEngine;
 
+    protected abstract void notifyParties(CaseData caseData);
+
     protected void sendNotification(Set<EmailDTO> recipients) {
         for (EmailDTO recipient : recipients) {
-            notificationService.sendMail(recipient.getTargetEmail(), recipient.getEmailTemplate(), recipient.getParameters(),
-                recipient.getReference());
+            notificationService.sendMail(
+                recipient.getTargetEmail(), recipient.getEmailTemplate(), recipient.getParameters(),
+                recipient.getReference()
+            );
         }
     }
-
-    public void notifyParties(CaseData caseData) {
-        Set<EmailDTO> partiesToEmail = new HashSet<>();
-        partiesToEmail.add(getApplicant(caseData));
-        partiesToEmail.addAll(getRespondents(caseData));
-        sendNotification(partiesToEmail);
-    }
-
-    protected abstract EmailDTO getApplicant(CaseData caseData);
-
-    protected abstract Set<EmailDTO> getRespondents(CaseData caseData);
 }
